significantly refactor the Corfu Object Proxy
