+++ src/main/java/org/corfudb/runtime/object/ObjectType.java
@@ -1,0 +1,9 @@
+package org.corfudb.runtime.object;
+
+/**
+ * Created by mwei on 3/30/16.
+ */
+public enum ObjectType {
+    SMR,
+    STATELESS
+}
