+++ src/main/java/org/corfudb/runtime/object/StaticMappingObject.java
@@ -2,6 +2,8 @@ package org.corfudb.runtime.object;
 
 /**
  * Created by mwei on 3/29/16.
+ *
+ * A dummy empty class for null SMR.
  */
 public class StaticMappingObject {
 }
