+++ src/main/java/org/corfudb/runtime/object/TransactionalMethod.java
@@ -1,6 +1,9 @@
 package org.corfudb.runtime.object;
 
 import java.lang.annotation.*;
+import java.util.Set;
+import java.util.UUID;
+import java.util.function.Function;
 
 /**
  * Created by mwei on 3/29/16.
@@ -9,4 +12,5 @@ import java.lang.annotation.*;
 @Retention(RetentionPolicy.RUNTIME)
 @Inherited
 public @interface TransactionalMethod {
+    boolean readOnly() default false;
 }
